refactor: Change the execute direcotry composition, and modify the Makefile with this change.
