+++ src/execute/dispatch/exec_disk.c
@@ -10,8 +10,8 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/minishell.h"
-#include "../../include/parser.h"
+#include "../../../include/minishell.h"
+#include "../../../include/parser.h"
 
 /*
 	Todo
