+++ src/execute/expansion/expand.c
@@ -1,22 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exec_subshell.c                                    :+:      :+:    :+:   */
+/*   expand.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tafujise <tafujise@student.42.jp>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/01/28 00:09:12 by tafujise          #+#    #+#             */
-/*   Updated: 2026/02/11 11:03:56 by tafujise         ###   ########.fr       */
+/*   Created: 2026/01/29 17:56:04 by tafujise          #+#    #+#             */
+/*   Updated: 2026/02/11 11:04:08 by tafujise         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/minishell.h"
-#include "../../include/parser.h"
+#include "../../../include/minishell.h"
+#include "../../../include/parser.h"
 
-t_status	exec_subshell(t_node *node, t_ctx *ctx, int pipe_in, int pipe_out)
+/*
+	Todo
+	- expand args and assignment word!!!.
+*/
+t_status	expand_words(t_simple_cmd *cmd, t_ctx *ctx)
 {
-	/*
-		Todo
-		- subshell
-	*/
 }
